YQL-19747 Move ranking.h
