+++ yql/essentials/sql/v1/complete/name/service/ranking/ranking.cpp
