+++ yql/essentials/sql/v1/complete/name/service/ranking/ranking.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <yql/essentials/sql/v1/complete/name/service/ranking/frequency.h>
+#include <yql/essentials/sql/v1/complete/name/service/name_service.h>
+
+#include <util/generic/hash.h>
+
+namespace NSQLComplete {
+
+    class IRanking: public TThrRefBase {
+    public:
+        using TPtr = TIntrusivePtr<IRanking>;
+
+        virtual void CropToSortedPrefix(TVector<TGenericName>& names, size_t limit) const = 0;
+        virtual ~IRanking() = default;
+    };
+
+    IRanking::TPtr MakeDefaultRanking();
+
+    IRanking::TPtr MakeDefaultRanking(TFrequencyData frequency);
+
+} // namespace NSQLComplete
