+++ yql/essentials/sql/v1/complete/name/service/static/ya.make
@@ -4,7 +4,6 @@ SRCS(
     json_name_set.cpp
     name_index.cpp
     name_service.cpp
-    ranking.cpp
 )
 
 PEERDIR(
